[Unit Tests] SparseTokensField (#19)
* add tests for SparseTokensField class; make BatchClusteringTaskTests.prepareKeyFieldInfo public


* Update SparseTokensFieldTests class; make BatchClusteringTaskTests.prepareXXXs public static


* Extract thos prepareXXX functions into a new testsPrepreUtils class; refactor SparseTokensFieldTests and BatchClusteringTaskTests


* refactor testsPrepareUtils


* update


* update


---------
