+++ src/test/java/org/opensearch/neuralsearch/sparse/algorithm/BatchClusteringTaskTests.java
@@ -4,51 +4,23 @@
  */
 package org.opensearch.neuralsearch.sparse.algorithm;
 
-import com.google.common.annotations.VisibleForTesting;
 import org.apache.commons.lang3.tuple.Pair;
-import org.apache.lucene.codecs.Codec;
-import org.apache.lucene.codecs.DocValuesProducer;
-import org.apache.lucene.codecs.FieldsProducer;
-import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.MergeState;
-import org.apache.lucene.index.SegmentInfo;
 import org.apache.lucene.index.FieldInfo;
-import org.apache.lucene.index.FieldInfos;
-import org.apache.lucene.index.IndexOptions;
-import org.apache.lucene.index.DocValuesType;
-import org.apache.lucene.index.DocValuesSkipIndexType;
-import org.apache.lucene.index.VectorEncoding;
-import org.apache.lucene.index.VectorSimilarityFunction;
-import org.apache.lucene.index.NumericDocValues;
-import org.apache.lucene.index.SortedDocValues;
-import org.apache.lucene.index.DocValuesSkipper;
-import org.apache.lucene.index.Terms;
-import org.apache.lucene.index.SortedSetDocValues;
-import org.apache.lucene.index.SortedNumericDocValues;
-import java.io.IOException;
-import org.apache.lucene.store.ByteBuffersDirectory;
 import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.InfoStream;
-import org.apache.lucene.util.Version;
-import org.apache.lucene.store.Directory;
 import org.junit.Before;
 import org.mockito.MockitoAnnotations;
 import org.opensearch.neuralsearch.sparse.AbstractSparseTestBase;
 import org.opensearch.neuralsearch.sparse.common.InMemoryKey;
+import org.opensearch.neuralsearch.sparse.testsPrepareUtils;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Iterator;
-import java.util.concurrent.Executors;
 import java.lang.reflect.Field;
 
 public class BatchClusteringTaskTests extends AbstractSparseTestBase {
-    @VisibleForTesting
     private List<BytesRef> terms;
     private InMemoryKey.IndexKey key;
-    private String fieldName = "test_field";
 
     @Before
     @Override
@@ -58,234 +30,6 @@ public class BatchClusteringTaskTests extends AbstractSparseTestBase {
 
         terms = Arrays.asList(new BytesRef("term1"), new BytesRef("term2"));
         key = new InMemoryKey.IndexKey(null, "test_field");
-    }
-
-    private FieldInfo prepareKeyFieldInfo() {
-
-        // Create a FieldInfo object
-        FieldInfo keyFieldInfo = new FieldInfo(
-            fieldName,                     // name
-            0,                             // number
-            false,                         // storeTermVector
-            false,                         // omitNorms
-            false,                         // storePayloads
-            IndexOptions.DOCS,             // indexOptions
-            DocValuesType.BINARY,          // docValuesType
-            DocValuesSkipIndexType.NONE,   // docValuesSkipIndex
-            -1,                            // dvGen
-            new HashMap<>(),               // attributes
-            0,                             // pointDimensionCount
-            0,                             // pointIndexDimensionCount
-            0,                             // pointNumBytes
-            0,                             // vectorDimension
-            VectorEncoding.FLOAT32,        // vectorEncoding
-            VectorSimilarityFunction.EUCLIDEAN, // vectorSimilarityFunction
-            false,                         // softDeletesField
-            false                          // isParentField
-        );
-        return keyFieldInfo;
-    }
-
-    private SegmentInfo prepareSegmentInfo() {
-        MergeState.DocMap[] docMaps = new MergeState.DocMap[1];
-        docMaps[0] = docID -> docID;
-        Directory dir = new ByteBuffersDirectory();
-        byte[] id = new byte[16];
-        for (int i = 0; i < id.length; i++) {
-            id[i] = (byte) i;
-        }
-
-        SegmentInfo segmentInfo = new SegmentInfo(
-            dir,                           // directory
-            Version.LATEST,                // version
-            Version.LATEST,                // minVersion
-            "_test_segment",               // name
-            10,                            // maxDoc
-            false,                         // isCompoundFile
-            false,                         // hasBlocks
-            Codec.getDefault(),            // codec
-            Collections.emptyMap(),        // diagnostics
-            id,                            // id
-            Collections.emptyMap(),        // attributes
-            null                           // indexSort
-        );
-        return segmentInfo;
-    }
-
-    private BinaryDocValues prepareBinaryDocValues() {
-        final BytesRef value = new BytesRef(new byte[] { 1, 2, 3, 4 });
-        BinaryDocValues binaryDocValues = new BinaryDocValues() {
-            private int docID = -1;
-
-            @Override
-            public int docID() {
-                return docID;
-            }
-
-            @Override
-            public int nextDoc() {
-                if (docID < 9) {
-                    docID++;
-                    return docID;
-                }
-                return NO_MORE_DOCS;
-            }
-
-            @Override
-            public int advance(int target) {
-                if (docID < target && target <= 9) {
-                    docID = target;
-                    return docID;
-                }
-                return NO_MORE_DOCS;
-            }
-
-            @Override
-            public long cost() {
-                return 10;
-            }
-
-            @Override
-            public BytesRef binaryValue() {
-                return value;
-            }
-
-            @Override
-            public boolean advanceExact(int target) throws IOException {
-                if (target <= 9) {
-                    docID = target;
-                    return true;
-                }
-                return false;
-            }
-        };
-        return binaryDocValues;
-    }
-
-    private DocValuesProducer prepareDocValuesProducer(BinaryDocValues binaryDocValues) {
-        DocValuesProducer docValuesProducer = new DocValuesProducer() {
-            @Override
-            public NumericDocValues getNumeric(FieldInfo field) {
-                return null;
-            }
-
-            @Override
-            public BinaryDocValues getBinary(FieldInfo field) {
-                if (field.name.equals(fieldName)) {
-                    return binaryDocValues;
-                }
-                return null;
-            }
-
-            @Override
-            public SortedDocValues getSorted(FieldInfo field) {
-                return null;
-            }
-
-            @Override
-            public SortedNumericDocValues getSortedNumeric(FieldInfo field) {
-                return null;
-            }
-
-            @Override
-            public SortedSetDocValues getSortedSet(FieldInfo field) {
-                return null;
-            }
-
-            @Override
-            public void checkIntegrity() {}
-
-            @Override
-            public void close() {}
-
-            @Override
-            public DocValuesSkipper getSkipper(FieldInfo field) throws IOException {
-                return null;
-            }
-        };
-        return docValuesProducer;
-    }
-
-    private FieldsProducer prepareFieldsProducer() {
-        FieldsProducer fieldsProducer = new FieldsProducer() {
-            @Override
-            public Iterator<String> iterator() {
-                return Collections.singleton(fieldName).iterator();
-            }
-
-            @Override
-            public Terms terms(String field) {
-                return null;
-            }
-
-            @Override
-            public int size() {
-                return 1;
-            }
-
-            @Override
-            public void checkIntegrity() {}
-
-            @Override
-            public void close() {}
-        };
-        return fieldsProducer;
-    }
-
-    private MergeState prepareMergeState(boolean isEmptyMaxDocs) {
-        MergeState.DocMap[] docMaps = new MergeState.DocMap[1];
-        docMaps[0] = docID -> docID;
-        SegmentInfo segmentInfo = prepareSegmentInfo();
-        // FieldInfo KeyFieldInfo = prepareKeyFieldInfo();
-
-        int[] maxDocs = new int[] { 10 };
-        if (isEmptyMaxDocs) {
-            maxDocs = new int[] { 0 };
-        }
-
-        // Create a FieldInfo object
-        FieldInfo keyFieldInfo = prepareKeyFieldInfo();
-
-        // Create a real BinaryDocValues object
-        BinaryDocValues binaryDocValues = prepareBinaryDocValues();
-
-        // Create a DocValuesProducer
-        DocValuesProducer docValuesProducer = prepareDocValuesProducer(binaryDocValues);
-
-        DocValuesProducer[] docValuesProducers = new DocValuesProducer[1];
-        docValuesProducers[0] = docValuesProducer;
-
-        // Create FieldInfos, like an array of FieldInfo
-        FieldInfos fieldInfos = new FieldInfos(new FieldInfo[] { keyFieldInfo });
-        FieldInfos[] fieldInfosArray = new FieldInfos[1];
-        fieldInfosArray[0] = fieldInfos;
-
-        // Create FieldsProducer
-        FieldsProducer fieldsProducer = prepareFieldsProducer();
-
-        FieldsProducer[] fieldsProducers = new FieldsProducer[1];
-        fieldsProducers[0] = fieldsProducer;
-
-        // Create MergeState
-        MergeState mergeState = new MergeState(
-            docMaps,
-            segmentInfo,
-            fieldInfos,                // mergeFieldInfos
-            null,                      // storedFieldsReaders
-            null,                      // termVectorsReaders
-            null,                      // normsProducers
-            docValuesProducers,        // docValuesProducers
-            fieldInfosArray,           // fieldInfos
-            null,                      // liveDocs
-            fieldsProducers,           // fieldsProducers
-            null,                      // pointsReaders
-            null,                      // knnVectorsReaders
-            maxDocs,
-            InfoStream.getDefault(),
-            Executors.newSingleThreadExecutor(),
-            false                      // needsIndexSort
-        );
-        return mergeState;
     }
 
     public void testConstructorDeepCopiesTerms() throws Exception {
@@ -346,8 +90,9 @@ public class BatchClusteringTaskTests extends AbstractSparseTestBase {
     public void testGetWithNonNullMergeState() throws Exception {
         // Test behavior with a not null merge state
         boolean isEmptyMaxDocs = false;
-        MergeState mergeState = prepareMergeState(isEmptyMaxDocs);
-        FieldInfo keyFieldInfo = prepareKeyFieldInfo();
+        testsPrepareUtils prepareHelper = new testsPrepareUtils();
+        MergeState mergeState = prepareHelper.prepareMergeState(isEmptyMaxDocs);
+        FieldInfo keyFieldInfo = prepareHelper.prepareKeyFieldInfo();
 
         // Create BatchClusteringTask
         BatchClusteringTask task = new BatchClusteringTask(terms, key, 0.5f, 0.3f, 10, mergeState, keyFieldInfo);
@@ -380,8 +125,9 @@ public class BatchClusteringTaskTests extends AbstractSparseTestBase {
     public void testGetWithNonNullMergeStateZeroMaxDocs() throws Exception {
         // Test behavior with a not null merge state
         boolean isEmptyMaxDocs = true;
-        MergeState mergeState = prepareMergeState(isEmptyMaxDocs);
-        FieldInfo keyFieldInfo = prepareKeyFieldInfo();
+        testsPrepareUtils prepareHelper = new testsPrepareUtils();
+        MergeState mergeState = prepareHelper.prepareMergeState(isEmptyMaxDocs);
+        FieldInfo keyFieldInfo = prepareHelper.prepareKeyFieldInfo();
 
         // Create BatchClusteringTask
         BatchClusteringTask task = new BatchClusteringTask(terms, key, 0.5f, 0.3f, 10, mergeState, keyFieldInfo);
